feat: implement role-based access control by adding role checks and removing permissions middleware

+++ src/modules/users/repositories/user.repository.js
@@ -3,7 +3,7 @@ import pool from "../../../config/db.config.js";
 export class UserRepository {
     async getAll() { 
         const result = await pool.query(
-            "SELECT * FROM users WHERE role = \'user\'"
+            "SELECT * FROM users WHERE role = $1", ['user']
         );
         // console.log('Result:', result.rows);
         return result.rows;
@@ -11,14 +11,14 @@ export class UserRepository {
 
     async getById(id) {
         const result = await pool.query(
-            "SELECT * FROM users WHERE id = $1", [id]
+            "SELECT * FROM users WHERE id = $1 AND role = $2", [id, 'user']
         );
         return result.rows[0];
     }
 
     async getByEmail(email) {
         const result = await pool.query(
-            "SELECT * FROM users WHERE email = $1", [email]
+            "SELECT * FROM users WHERE email = $1 AND role = $2", [email, 'user']
         );
         return result.rows[0];
     }
@@ -51,27 +51,28 @@ export class UserRepository {
         values.push(id);
 
         const result = await pool.query(
-            `UPDATE users SET ${fields.join(', ')} WHERE id = $${counter} RETURNING *`, values
+            `UPDATE users SET ${fields.join(', ')} WHERE id = $${counter} AND role = 'user' RETURNING *`, values
         );
         return result.rows[0];
     }
 
     async delete(id) {
         const result = await pool.query(
-            "DELETE FROM users WHERE id = $1 RETURNING *", [id]
+            "DELETE FROM users WHERE id = $1 AND role = $2 RETURNING *", [id, 'user']
         );
+        return result.rows[0];
     }
 
     async activateUser(id) {
         const result = await pool.query(
-            "UPDATE users SET is_active = true WHERE id = $1 RETURNING *", [id]
+            "UPDATE users SET is_active = true WHERE id = $1 AND role = $2 RETURNING *", [id, 'user']
         );
         return result.rows[0];
     }
 
     async updatePassword(id, password) {
         const result = await pool.query(
-            "UPDATE users SET password = $1 WHERE id = $2 RETURNING *", [password, id]
+            "UPDATE users SET password = $1 WHERE id = $2 AND role = $3 RETURNING *", [password, id, 'role']
         );
         if(result.rowCount === 0 || !result.rows[0]) {
             throw new Error('User not found');
@@ -81,7 +82,7 @@ export class UserRepository {
 
     async deactivateUser(id) {
         const result = await pool.query(
-            'UPDATE users SET is_active = false WHERE id = $1 RETURNING *', [id]
+            'UPDATE users SET is_active = false WHERE id = $1 AND role = $2 RETURNING *', [id, 'user']
         );
         return result.rows[0];
     }
